refactor: app move some to auth
